Move sql source base class to source package
